backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
duckdb>=1.1,<2.0
python-dotenv>=1.0,<2.0
openai>=1.100,<2.0
faiss-cpu>=1.8,<2.0
text2vec>=1.3,<2.0
torch>=2.2,<3.0
